+++ packages/open-payments/src/client/outgoing-payment.ts
@@ -30,7 +30,7 @@ export interface OutgoingPaymentRoutes {
 export const createOutgoingPaymentRoutes = (
   deps: RouteDeps
 ): OutgoingPaymentRoutes => {
-  const { axiosInstance, openApi, logger } = deps
+  const { openApi, ...baseDeps } = deps
 
   const getOutgoingPaymentOpenApiValidator =
     openApi.createResponseValidator<OutgoingPayment>({
@@ -53,13 +53,13 @@ export const createOutgoingPaymentRoutes = (
   return {
     get: (requestArgs: ResourceRequestArgs) =>
       getOutgoingPayment(
-        { axiosInstance, logger },
+        baseDeps,
         requestArgs,
         getOutgoingPaymentOpenApiValidator
       ),
     list: (requestArgs: CollectionRequestArgs, pagination?: PaginationArgs) =>
       listOutgoingPayments(
-        { axiosInstance, logger },
+        baseDeps,
         requestArgs,
         listOutgoingPaymentOpenApiValidator,
         pagination
@@ -69,7 +69,7 @@ export const createOutgoingPaymentRoutes = (
       createArgs: CreateOutgoingPaymentArgs
     ) =>
       createOutgoingPayment(
-        { axiosInstance, logger },
+        baseDeps,
         requestArgs,
         createOutgoingPaymentOpenApiValidator,
         createArgs
@@ -82,11 +82,10 @@ export const getOutgoingPayment = async (
   requestArgs: ResourceRequestArgs,
   validateOpenApiResponse: ResponseValidator<OutgoingPayment>
 ) => {
-  const { axiosInstance, logger } = deps
   const { url, accessToken } = requestArgs
 
   const outgoingPayment = await get(
-    { axiosInstance, logger },
+    deps,
     {
       url,
       accessToken
@@ -112,12 +111,11 @@ export const createOutgoingPayment = async (
   validateOpenApiResponse: ResponseValidator<OutgoingPayment>,
   createArgs: CreateOutgoingPaymentArgs
 ) => {
-  const { axiosInstance, logger } = deps
   const { url: baseUrl, accessToken } = requestArgs
   const url = `${baseUrl}${getRSPath('/outgoing-payments')}`
 
   const outgoingPayment = await post(
-    { axiosInstance, logger },
+    deps,
     { url, body: createArgs, accessToken },
     validateOpenApiResponse
   )
@@ -140,12 +138,11 @@ export const listOutgoingPayments = async (
   validateOpenApiResponse: ResponseValidator<OutgoingPaymentPaginationResult>,
   pagination?: PaginationArgs
 ) => {
-  const { axiosInstance, logger } = deps
   const { url: baseUrl, accessToken, walletAddress } = requestArgs
   const url = `${baseUrl}${getRSPath('/outgoing-payments')}`
 
   const outgoingPayments = await get(
-    { axiosInstance, logger },
+    deps,
     {
       url,
       accessToken,