+++ packages/open-payments/src/client/wallet-address.ts
@@ -1,43 +1,54 @@
 import { HttpMethod } from '@interledger/openapi'
 import { RouteDeps, UnauthenticatedResourceRequestArgs } from '.'
-import { JWKS, WalletAddress, getRSPath } from '../types'
+import { JWKS, WalletAddress, DIDDocument, getWAPath } from '../types'
 import { get } from './requests'
 
 export interface WalletAddressRoutes {
   get(args: UnauthenticatedResourceRequestArgs): Promise<WalletAddress>
   getKeys(args: UnauthenticatedResourceRequestArgs): Promise<JWKS>
+  // TODO: Define schema for DID Document
+  getDIDDocument(args: UnauthenticatedResourceRequestArgs): Promise<DIDDocument>
 }
 
 export const createWalletAddressRoutes = (
   deps: RouteDeps
 ): WalletAddressRoutes => {
-  const { axiosInstance, openApi, logger } = deps
+  const { openApi, ...baseDeps } = deps
 
-  const getPaymentPaymentValidator =
+  const getWalletAddressValidator =
     openApi.createResponseValidator<WalletAddress>({
-      path: getRSPath('/'),
+      path: getWAPath('/'),
       method: HttpMethod.GET
     })
 
-  const getPaymentPaymentKeysValidator = openApi.createResponseValidator<JWKS>({
-    path: getRSPath('/jwks.json'),
+  const getWalletAddressKeysValidator = openApi.createResponseValidator<JWKS>({
+    path: getWAPath('/jwks.json'),
+    method: HttpMethod.GET
+  })
+
+  const getDidDocumentValidator = openApi.createResponseValidator<DIDDocument>({
+    path: getWAPath('/did.json'),
     method: HttpMethod.GET
   })
 
   return {
     get: (args: UnauthenticatedResourceRequestArgs) =>
-      get(
-        { axiosInstance, logger },
-        { url: args.url },
-        getPaymentPaymentValidator
-      ),
+      get(baseDeps, args, getWalletAddressValidator),
     getKeys: (args: UnauthenticatedResourceRequestArgs) =>
       get(
-        { axiosInstance, logger },
+        baseDeps,
         {
           url: `${args.url}/jwks.json`
         },
-        getPaymentPaymentKeysValidator
+        getWalletAddressKeysValidator
+      ),
+    getDIDDocument: (args: UnauthenticatedResourceRequestArgs) =>
+      get(
+        baseDeps,
+        {
+          url: `${args.url}/did.json`
+        },
+        getDidDocumentValidator
       )
   }
 }