+++ packages/open-payments/src/test/helpers.ts
@@ -16,11 +16,13 @@ import {
   Grant,
   IncomingPaymentWithPaymentMethods,
   IlpPaymentMethod,
-  PublicIncomingPayment
+  PublicIncomingPayment,
+  DIDDocument
 } from '../types'
 import { v4 as uuid } from 'uuid'
 import { ResponseValidator } from '@interledger/openapi'
 import base64url from 'base64url'
+import { BaseDeps } from '../client'
 
 export const silentLogger = createLogger({
   level: 'silent'
@@ -56,11 +58,19 @@ export const mockJwk = (overrides?: Partial<JWK>): JWK => ({
   ...overrides
 })
 
+export const mockDIDDocument = (
+  overrides?: Partial<DIDDocument>
+): DIDDocument => ({
+  // TODO - Flesh this out when we have a more detailed DID Document spec
+  ...overrides
+})
+
 export const mockWalletAddress = (
   overrides?: Partial<WalletAddress>
 ): WalletAddress => ({
   id: 'https://example.com/.well-known/pay',
   authServer: 'https://auth.wallet.example/authorize',
+  resourceServer: 'https://wallet.example/op',
   assetScale: 2,
   assetCode: 'USD',
   ...overrides
@@ -301,5 +311,12 @@ export const mockQuote = (overrides?: Partial<Quote>): Quote => ({
   createdAt: new Date().toISOString(),
   expiresAt: new Date(Date.now() + 60_000).toISOString(),
   method: 'ilp',
+  ...overrides
+})
+
+export const createTestDeps = (overrides?: Partial<BaseDeps>): BaseDeps => ({
+  axiosInstance: defaultAxiosInstance,
+  logger: silentLogger,
+  useHttp: false,
   ...overrides
 })